flowork-core/requirements.txt
# ===================================================================
# FLOWORK CORE - REQUIREMENTS (VERSI 1.3 - OPEN CORE MVP + FIXES)
# ===================================================================
--extra-index-url https://download.pytorch.org/whl/cu121
torch>=2.5.1,<3.0.0
torchvision>=0.20.1,<0.21.0
torchaudio>=2.5.1,<3.0.0
llama-cpp-python>=0.3.14,<0.4.0
sentence-transformers>=5.0.0,<6.0.0
transformers>=4.54.1,<5.0.0
diffusers
accelerate
safetensors>=0.5.3,<0.6.0
hf-xet>=1.1.5,<2.0.0
google-generativeai>=0.8.5,<0.9.0
faster-whisper
openai
pyaudio
mss
moviepy
pystray
Pillow>=10.0.0
pandas>=2.3.1,<3.0.0
scipy<1.11
beautifulsoup4>=4.13.4,<5.0.0
selenium>=4.34.2,<5.0.0
webdriver-manager>=4.0.2,<5.0.0
fake-useragent>=2.2.0,<3.0.0
duckduckgo-search>=8.1.1,<9.0.0
eventlet
python-socketio==5.12.1
psutil>=7.0.0,<8.0.0
GPUtil>=1.4.0,<2.0.0
requests>=2.32.4,<3.0.0
pydantic>=2.11.7,<3.0.0
apscheduler==3.10.4
python-dotenv>=1.1.1,<2.0.0
flask
flask-cors
watchdog>=6.0.0,<7.0.0
pycryptodome>=3.23.0,<4.0.0
cryptography>=45.0.5,<46.0.0
python-dateutil>=2.9.0.post0,<3.0.0
cython>=3.1.2,<4.0.0
toml>=0.10.2,<0.11.0
packaging
msgpack
prometheus-client
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
aiohttp
aiofiles
websockets==11.0.3
web3
eth-account
prometheus-flask-exporter
werkzeug
SQLAlchemy
flask-migrate
